Module aws_sdk_resourcegroupstagging::error

Modules

concurrent_modification_exception

See ConcurrentModificationException

constraint_violation_exception

See ConstraintViolationException

internal_service_exception

See InternalServiceException

invalid_parameter_exception

See InvalidParameterException

pagination_token_expired_exception

See PaginationTokenExpiredException

throttled_exception

See ThrottledException

Structs

ConcurrentModificationException

The target of the operation is currently being modified by a different request. Try

ConstraintViolationException

The request was denied because performing this operation violates a constraint.

DescribeReportCreationError

GetComplianceSummaryError

GetResourcesError

GetTagKeysError

GetTagValuesError

InternalServiceException

The request processing failed because of an unknown error, exception, or failure. You

InvalidParameterException

This error indicates one of the following:

PaginationTokenExpiredException

A PaginationToken is valid for a maximum of 15 minutes. Your request was

StartReportCreationError

TagResourcesError

ThrottledException

The request was denied to limit the frequency of submitted requests.

UntagResourcesError

Enums

DescribeReportCreationErrorKind

GetComplianceSummaryErrorKind

GetResourcesErrorKind

GetTagKeysErrorKind

GetTagValuesErrorKind

StartReportCreationErrorKind

TagResourcesErrorKind

UntagResourcesErrorKind