Module aws_sdk_resourcegroupstagging::model

Modules

compliance_details

See ComplianceDetails

failure_info

See FailureInfo

resource_tag_mapping

See ResourceTagMapping

summary

See Summary

tag

See Tag

tag_filter

See TagFilter

Structs

ComplianceDetails

Information that shows whether a resource is compliant with the effective tag policy,

FailureInfo

Information about the errors that are returned for each failed resource. This

ResourceTagMapping

A list of resource ARNs and the tags (keys and values) that are associated with

Summary

A count of noncompliant resources.

Tag

The metadata that you apply to AWS resources to help you categorize and organize

TagFilter

A list of tags (keys and values) that are used to specify the associated

Enums

ErrorCode

GroupByAttribute

TargetIdType